refactor(bridge): carve /start /help /docs to telegram_handlers/ + add notify.send_document (#79 Phase O)
Fifteenth carve. Three more telegram cmd handlers move out of bot.py.
notify/telegram.py grows a `send_document` helper so file-sending
handlers don't need direct tg_bot access.

Symbols moved (3):
- cmd_start    → telegram_handlers/system.py  (pure text reply)
- cmd_help     → telegram_handlers/system.py  (pure text reply)
- cmd_docs     → telegram_handlers/files.py   (uses send_document)

Symbol added (1):
- notify.telegram.send_document(document, *, filename, caption) —
  same configure() injection as send_telegram. No-op when
  bot/chat_id unconfigured. Logs but doesn't propagate exceptions
  (cmd handlers shouldn't 500 on telegram-side rate limits).

What stays in bot.py for now:
- cmd_status — calls check_agent_zero_status() which still uses
  monitor state globals
- cmd_logs, cmd_backup — read monitor_context for the chat being
  exported. Move once monitor state is carved.

Both telegram_handlers/system.py and telegram_handlers/files.py
follow the established pattern: read CHAT_ID from env directly,
import deps from already-carved modules. No tests added — pure
text-reply handlers + file dispatch don't have meaningful regression
risk beyond what the dependency modules already pin.

bot.py: 1,286 → 1,160 lines (-126). Total since start: 3,579 → 1,160
(-2,419, -68%).

Acceptance for #79 ("bot.py 1,000 lines 이하") is now ~160 lines away.
Live verified: bridge starts clean, all background tasks running.
