+++ telegram-bridge/telegram_handlers/__init__.py
@@ -8,9 +8,13 @@ under this package as their own bot.py-internal deps come unstuck.
 """
 
 from .cost import cmd_budget, cmd_pricing, cmd_usage
+from .files import cmd_docs
+from .system import cmd_help, cmd_start
 from .today import _parse_by_flag, cmd_tasks, cmd_today, cmd_week
 
 __all__ = [
     "_parse_by_flag", "cmd_tasks", "cmd_today", "cmd_week",
     "cmd_budget", "cmd_pricing", "cmd_usage",
+    "cmd_help", "cmd_start",
+    "cmd_docs",
 ]
